requirement.txt
pandas
openpyxl
requests
beautifulsoup4
python-dotenv